Initial separation of gui_final into MVC files
