--- config.py
@@ -1,4 +1,0 @@
-import configparser
-
-config = configparser.RawConfigParser()
-config.read('config.ini')
