+++ networking/server_info.py
@@ -3,7 +3,6 @@
 # Information on the values of headers sent:
 import csv
 import struct
-from config import config
 
 
 class ServerInfo:
@@ -52,22 +51,6 @@ class ServerInfo:
         TC2_SEND: 'TC2',
         TC3_SEND: 'TC3'
     }
-
-    c = config.get("Calibration", "LC1_SEND").split(",")
-    print(tuple(c))
-
-    # calibrations = {
-    #     LC1_SEND: (1, 0),
-    #     LC_MAIN_SEND: (0.1365, -66.885),
-    #     LC2_SEND: (1, 0),
-    #     LC3_SEND: (1, 0),
-    #     PT_FEED_SEND: (-0.275787487, 1069),
-    #     PT_COMB_SEND: (-0.2810327855, 1068),
-    #     PT_INJE_SEND: (-0.2782331275, 1045),
-    #     TC1_SEND: (0.1611, -250),
-    #     TC2_SEND: (0.1611, -250),
-    #     TC3_SEND: (0.1611, -250)
-    # }
 
     calibrations = {
         LC_MAIN_SEND: (-.03159, 105),
